Core: Split KiCAD sch file from fileformats
